JNI: fixing verifyCoherency type
Just noticed that this "long" is an "int" in C++, and the value we use
to set it is also an int.

Bug: 331807348
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5601300
Auto-Submit: Sam Maier <smaier@chromium.org>
Commit-Queue: Tomasz Wiszkowski <ender@google.com>
Cr-Commit-Position: refs/heads/main@{#1310965}

+++ components/omnibox/browser/android/java/src/org/chromium/components/omnibox/AutocompleteResult.java
@@ -218,6 +218,6 @@ public class AutocompleteResult {
     @NativeMethods
     interface Natives {
         boolean verifyCoherency(
-                long nativeAutocompleteResult, long[] matches, long suggestionIndex, int origin);
+                long nativeAutocompleteResult, long[] matches, int suggestionIndex, int origin);
     }
 }
